+++ lib/fortnite.dart
@@ -23,6 +23,7 @@ export "src/structures/athena_cosmetic.dart";
 export "src/structures/cosmetic_locker.dart";
 export "src/structures/player.dart";
 export "src/structures/banner_quest.dart";
+export "src/structures/survivor_squad_preset.dart";
 
 export "src/util/utils.dart";
 export "src/util/curve_table.dart";