+++ lib/src/client/client.dart
@@ -212,6 +212,30 @@ class Client {
     }
   }
 
+  /// send a get request to epic games.
+  ///
+  /// get request to [url].
+  Future<dynamic> get(
+    String url, {
+    String? overrideToken,
+  }) async =>
+      await send(method: "GET", url: url, overrideToken: overrideToken);
+
+  /// send a post request to epic games.
+  ///
+  /// post request to [url] with [body].
+  Future<dynamic> post(
+    String url, {
+    dynamic body,
+    String? overrideToken,
+  }) async =>
+      await send(
+        method: "POST",
+        url: url,
+        body: body,
+        overrideToken: overrideToken,
+      );
+
   /// returns the avatar of the given account ids
   Future<List<Avatar>> getAvatars(List<String> accountIds) async {
     List<Avatar> avatars = [];